finish track content ui
